Update the example plugin to copy the template
The example plugin should be a better copy of the template (eg use two files,
connector.go and plugin.go) so that it's clearer to people trying to build
plugins exactly how to create and build them in practice

This reorganizes the content to better mimic the templates and updates
the Dockerfile to correctly build the plugin.
