+++ test/plugin/Dockerfile
@@ -7,7 +7,7 @@ COPY . .
 RUN mkdir -p /usr/local/lib/secretless && \
     go build -buildmode=plugin \
              -o /usr/local/lib/secretless/example-plugin.so \
-                 ./example/example_plugin.go
+                 ./example/
 
 # Do not remove this - we are intentionally trying to exercise
 # limited user functionality
